requirements/local.txt
#
# This file is autogenerated by pip-compile with Python 3.9
# by the following command:
#
#    pip-compile requirements/local.in
#
alembic==1.11.1
    # via flask-migrate
aniso8601==9.0.1
    # via flask-restful
blinker==1.6.2
    # via flask
click==8.1.3
    # via flask
flask==2.3.2
    # via
    #   -r requirements/base.in
    #   flask-cors
    #   flask-marshmallow
    #   flask-migrate
    #   flask-restful
    #   flask-sqlalchemy
flask-cors==4.0.0
    # via -r requirements/base.in
flask-marshmallow==0.15.0
    # via -r requirements/base.in
flask-migrate==4.0.4
    # via -r requirements/base.in
flask-restful==0.3.10
    # via -r requirements/base.in
flask-sqlalchemy==3.0.5
    # via
    #   -r requirements/base.in
    #   flask-migrate
importlib-metadata==6.7.0
    # via flask
itsdangerous==2.1.2
    # via flask
jinja2==3.1.2
    # via flask
mako==1.2.4
    # via alembic
markupsafe==2.1.3
    # via
    #   jinja2
    #   mako
    #   werkzeug
marshmallow==3.19.0
    # via
    #   flask-marshmallow
    #   marshmallow-sqlalchemy
marshmallow-sqlalchemy==0.29.0
    # via -r requirements/base.in
packaging==23.1
    # via
    #   flask-marshmallow
    #   marshmallow
    #   marshmallow-sqlalchemy
psycopg2==2.9.6
    # via -r requirements/base.in
python-dotenv==1.0.0
    # via -r requirements/base.in
pytz==2023.3
    # via flask-restful
six==1.16.0
    # via flask-restful
sqlalchemy==2.0.18
    # via
    #   alembic
    #   flask-sqlalchemy
    #   marshmallow-sqlalchemy
typing-extensions==4.7.1
    # via
    #   alembic
    #   sqlalchemy
werkzeug==2.3.6
    # via flask
zipp==3.15.0
    # via importlib-metadata
